break out extension code to src/extension-provider
